# optimize scope delegation (Act 3)