Retake on https://github.com/sbt/sbt/pull/4003

Using `time sbt exit && time sbt exit && time sbt exit` as the test:

| build                               | baseline (b82a187) | this PR (3436891) |
|-------------------------------------|--------------------|-------------------|
| Cinnamon (e531517, 144474 settings) | 39.633s            | 36.081s           |
| Akka (4cbc049, 31875 settings)      | 17.133s            | 16.128            |